move dub geojson code -> shared geojson code base
