+++ gpkg/getfeaturesprovider.go
@@ -9,7 +9,7 @@ import (
 )
 
 type GetFeaturesProvider struct {
-	data        FeatureCollectionGeoJSON
+	data        provider.FeatureCollectionGeoJSON
 	srsid       string
 	contenttype string
 }
