+++ postgis/getfeatureprovider.go
@@ -9,7 +9,7 @@ import (
 )
 
 type GetFeatureProvider struct {
-	data        *Feature
+	data        *provider.Feature
 	srsid       string
 	contenttype string
 }
