+++ postgis/postgis.go
@@ -59,8 +59,8 @@ func (postgis Postgis) Close() error {
 	return postgis.db.Close()
 }
 
-func (postgis Postgis) GetFeatures(ctx context.Context, db *sqlx.DB, collection provider.Collection, whereMap map[string]string, offset uint64, limit uint64, featureId interface{}, bbox [4]float64) (result FeatureCollectionGeoJSON, err error) {
-	result = FeatureCollectionGeoJSON{}
+func (postgis Postgis) GetFeatures(ctx context.Context, db *sqlx.DB, collection provider.Collection, whereMap map[string]string, offset uint64, limit uint64, featureId interface{}, bbox [4]float64) (result provider.FeatureCollectionGeoJSON, err error) {
+	result = provider.FeatureCollectionGeoJSON{}
 	if len(bbox) > 4 {
 		err = errors.New("bbox with 6 elements not supported")
 		return
@@ -140,7 +140,7 @@ func (postgis Postgis) GetFeatures(ctx context.Context, db *sqlx.DB, collection 
 
 	result.NumberReturned = 0
 	result.Type = "FeatureCollection"
-	result.Features = make([]*Feature, 0)
+	result.Features = make([]*provider.Feature, 0)
 
 	for rows.Next() {
 		if err = ctx.Err(); err != nil {
@@ -160,7 +160,7 @@ func (postgis Postgis) GetFeatures(ctx context.Context, db *sqlx.DB, collection 
 			return
 		}
 
-		feature := &Feature{Type: "Feature", Properties: make(map[string]interface{})}
+		feature := &provider.Feature{Feature: geojson.Feature{Properties: make(map[string]interface{})}}
 
 		for i, colName := range cols {
 			// check if the context cancelled or timed out
