+++ provider/geojson.go
@@ -1,13 +1,10 @@
-package postgis
+package provider
 
 import (
 	"oaf-server/codegen"
 
 	"github.com/go-spatial/geom/encoding/geojson"
 )
-
-// this code is generated by go generate
-// Not anymore geojson.FeatureType included
 
 type FeatureCollectionGeoJSON struct {
 	NumberReturned int64          `json:"numberReturned,omitempty"`
@@ -16,16 +13,14 @@ type FeatureCollectionGeoJSON struct {
 	Features       []*Feature     `json:"features"`
 	Links          []codegen.Link `json:"links,omitempty"`
 	NumberMatched  int64          `json:"numberMatched,omitempty"`
-	Offset         int64
+	Crs            string         `json:"crs,omitempty"`
+	Offset         int64          `json:"-"`
 }
 
 type Feature struct {
-	Type string      `json:"type"`
-	ID   interface{} `json:"id,omitempty"`
-	// can be null
-	Geometry geojson.Geometry `json:"geometry"`
-	// can be null
-	Properties map[string]interface{} `json:"properties"`
-	// can be null
+	// overwrite ID in geojson.Feature so strings are also allowed as id
+	ID interface{} `json:"id,omitempty"`
+	geojson.Feature
+	// Added Links in de document
 	Links []codegen.Link `json:"links,omitempty"`
 }
